Removed helper module
