+++ apps/common/lib/lexical/ast/detection/function_capture.ex
@@ -4,7 +4,7 @@ defmodule Lexical.Ast.Detection.FunctionCapture do
   alias Lexical.Document
   alias Lexical.Document.Position
 
-  @behaviour Detection
+  use Detection
 
   @impl Detection
   def detected?(%Document{} = document, %Position{} = position) do
