+++ apps/common/lib/lexical/ast/detection/pipe.ex
@@ -4,7 +4,7 @@ defmodule Lexical.Ast.Detection.Pipe do
   alias Lexical.Document
   alias Lexical.Document.Position
 
-  @behaviour Detection
+  use Detection
 
   @impl Detection
   def detected?(%Document{} = document, %Position{} = position) do
