+++ apps/common/lib/lexical/ast/detection/spec.ex
@@ -1,13 +1,12 @@
 defmodule Lexical.Ast.Detection.Spec do
   alias Lexical.Ast.Detection
-  alias Lexical.Ast.Detection.Ancestor
   alias Lexical.Document
   alias Lexical.Document.Position
 
-  @behaviour Detection
+  use Detection
 
   @impl Detection
   def detected?(%Document{} = document, %Position{} = position) do
-    Ancestor.is_spec?(document, position)
+    ancestor_is_spec?(document, position)
   end
 end
